# Event/docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Event API",
    "version": "1.0.0",
    "description": "This API allows users to manage events and address events.",
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "/event/api"
    }
  ],
  "paths": {
...
    "/events/{id}/status": {
      "put": {
        "summary": "Approve a specific event by ID",
        "description": "Allows an admin to approve an event by updating its status to 'Aprovado'. The event must have a current status of 'Pendente'.",
        "operationId": "updateEventStatus",
        "tags": [
          "Events"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "The unique identifier of the event to approve.",
            "schema": {
              "type": "string",
              "example": "123e4567-e89b-12d3-a456-426614174000"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Event status updated successfully to 'Approved'.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Event"
                }
              }
            }
          },
          "400": {
            "description": "Event status is not 'Pending'.",
            "content": {
              "application/json": {
                "examples": {
                  "notPending": {
                    "summary": "Invalid status",
                    "value": {
                      "message": "Event status is not pending"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Event or required status not found.",
            "content": {
              "application/json": {
                "examples": {
                  "eventNotFound": {
                    "summary": "Event not found",
                    "value": {
                      "message": "Event not found"
                    }
                  },
                  "statusNotFound": {
                    "summary": "Status not found",
                    "value": {
                      "message": "Event status not found"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error - An unexpected error occurred while updating the event status.",
            "content": {
              "application/json": {
                "examples": {
                  "serverError": {
                    "summary": "Unexpected error",
                    "value": {
                      "message": "Error updating event status"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/eventStatus": {
      "get": {
        "summary": "Get all event statuses",
        "description": "Retrieve a list of all event statuses available in the system.",
        "operationId": "getAllEventStatuses",
        "tags": [
          "EventStatus"
        ],
        "responses": {
          "200": {
            "description": "A list of event statuses",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EventStatus"
                  }
                },
                "example": [
                  {
                    "eventStatusID": "11111111-1111-1111-1111-111111111111",
                    "status": "Scheduled"
                  },
                  {
                    "eventStatusID": "22222222-2222-2222-2222-222222222222",
                    "status": "Completed"
                  }
                ]
              }
            }
          },
          "404": {
            "description": "No event statuses found"
          }
        }
      },
      "post": {
        "summary": "Create a new event status",
        "description": "Create a new event status in the system with the provided status value.",
        "operationId": "createEventStatus",
        "tags": [
          "EventStatus"
        ],
        "requestBody": {
          "description": "Data for creating a new event status",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventStatusInput"
              },
              "example": {
                "status": "In Progress"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Event status created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventStatus"
                },
                "example": {
                  "eventStatusID": "33333333-3333-3333-3333-333333333333",
                  "status": "In Progress"
                }
              }
            }
          },
          "400": {
            "description": "Missing required field: status"
          },
          "500": {
            "description": "Error creating event status"
          }
        }
      }
    },
    "/eventStatus/{id}": {
      "get": {
        "summary": "Get an event status by ID",
        "description": "Retrieve the details of a specific event status using its unique identifier.",
        "operationId": "getEventStatusById",
        "tags": [
          "EventStatus"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "The ID of the event status",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The event status object",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventStatus"
                },
                "example": {
                  "eventStatusID": "44444444-4444-4444-4444-444444444444",
                  "status": "Cancelled"
                }
              }
            }
          },
          "404": {
            "description": "Event status not found"
          },
          "500": {
            "description": "Error fetching event status"
          }
        }
      },
      "put": {
        "summary": "Update an event status",
        "description": "Update the status of an existing event using the provided ID and new status data.",
        "operationId": "updateEventStatus",
        "tags": [
          "EventStatus"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "The ID of the event status to update",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Updated event status data",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventStatusInput"
              },
              "example": {
                "status": "Postponed"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Event status updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventStatus"
                },
                "example": {
                  "eventStatusID": "55555555-5555-5555-5555-555555555555",
                  "status": "Postponed"
                }
              }
            }
          },
          "404": {
            "description": "Event status not found"
          },
          "500": {
            "description": "Error updating event status"
          }
        }
      },
      "delete": {
        "summary": "Delete an event status",
        "description": "Delete a specific event status from the system using its ID.",
        "operationId": "deleteEventStatus",
        "tags": [
          "EventStatus"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "The ID of the event status to delete",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Event status deleted successfully"
          },
          "404": {
            "description": "Event status not found"
          },
          "500": {
            "description": "Error deleting event status"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "apiKey",
        "name": "Authorization",
        "in": "header",
        "description": "Bearer token authentication"
      }
    },
    "schemas": {
      "Event": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "ID of the event"
          },
          "name": {
            "type": "string",
            "description": "Name of the event"
          },
          "description": {
            "type": "string",
            "description": "Description of the event"
          },
          "startAt": {
            "type": "string",
            "format": "date-time",
            "description": "Start date and time of the event"
          },
          "endAt": {
            "type": "string",
            "format": "date-time",
            "description": "End date and time of the event"
          },
          "price": {
            "type": "number",
            "description": "Price of the event"
          },
          "eventPicture": {
            "type": "string",
            "description": "Event status ID"
          }
        }
      },
...
      "EventStatus": {
        "type": "object",
        "properties": {
          "eventStatusID": {
            "type": "string",
            "example": "11111111-1111-1111-1111-111111111111"
          },
          "status": {
            "type": "string",
            "example": "Scheduled"
          }
        }
      },
      "EventStatusInput": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "Completed"
          }
        },
        "required": [
          "status"
        ]
      }
    }
  }
}